cli/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "epd-tool"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "Pillow>=10.0",
    "Jinja2>=3.0",
    "PyYAML>=6.0",
    "requests>=2.28",
    "zeroconf>=0.131",
]

[project.scripts]
epd = "epd_tool.cli:main"
